fix(cms): make the branch start again
Two independent faults kept Django from booting at all, so nothing on this
branch could be exercised - which is also why CI stayed green.

- `INSTALLED_APPS` listed `djangocms_versioning`, `djangocms_alias` and
  `djangocms_stories` twice: once in the main list and once in a trailing
  `INSTALLED_APPS +=` block left over from the 4.1 stage. Django refused to
  start with `ImproperlyConfigured: Application labels aren't unique`.
  The main list already orders them correctly, so the trailing block goes.
  `CMS_CONFIRM_VERSION4` stays - django-cms 5.0 still requires it.

- The story post admin lived in `vitrina/users/admin.py`, which imports
  `djangocms_stories.admin` while `vitrina.users` is still loading. That
  triggers cms plugin discovery mid-import and fails with a circular
  `ImportError` on `SortedManyToManyAutocompleteMixin`. Moved to
  `vitrina/cms/admin.py`: `vitrina.users` must load before `cms`, and this
  admin must load after `djangocms_stories.admin` - `vitrina.cms` sits after
  `djangocms_stories` in INSTALLED_APPS, so both constraints hold.

`manage.py check` passes again.
